report_generation/requirements.txt
pandas
plotly
sqlalchemy
numpy
python-dotenv
typing
boto3
kaleido